packages/happtiq_commons_google_cloud/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile pyproject.toml
#
--extra-index-url https://europe-west1-python.pkg.dev/happtiq-mabula-demo-play/quickstart-python-repo/simple/

cachetools==5.4.0
    # via google-auth
certifi==2024.7.4
    # via requests
charset-normalizer==3.3.2
    # via requests
cloudevents==1.11.0
    # via python_utils_happtiq (pyproject.toml)
deprecation==2.1.0
    # via cloudevents
google-api-core[grpc]==2.19.1
    # via
    #   google-cloud-appengine-logging
    #   google-cloud-core
    #   google-cloud-logging
    #   google-cloud-pubsub
    #   google-cloud-secret-manager
    #   google-cloud-storage
google-auth==2.32.0
    # via
    #   google-api-core
    #   google-auth-oauthlib
    #   google-cloud-appengine-logging
    #   google-cloud-core
    #   google-cloud-pubsub
    #   google-cloud-secret-manager
    #   google-cloud-storage
    #   python_utils_happtiq (pyproject.toml)
google-auth-oauthlib==1.2.1
    # via python_utils_happtiq (pyproject.toml)
google-cloud-appengine-logging==1.4.4
    # via google-cloud-logging
google-cloud-audit-log==0.2.0
    # via google-cloud-logging
google-cloud-core==2.4.1
    # via
    #   google-cloud-logging
    #   google-cloud-storage
google-cloud-logging==3.1.1
    # via python_utils_happtiq (pyproject.toml)
google-cloud-pubsub==2.22.0
    # via python_utils_happtiq (pyproject.toml)
google-cloud-secret-manager==2.20.1
    # via python_utils_happtiq (pyproject.toml)
google-cloud-storage==2.18.0
    # via python_utils_happtiq (pyproject.toml)
google-crc32c==1.5.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.1
    # via google-cloud-storage
googleapis-common-protos[grpc]==1.63.2
    # via
    #   google-api-core
    #   google-cloud-audit-log
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.12.4
    # via
    #   google-cloud-logging
    #   google-cloud-pubsub
    #   google-cloud-secret-manager
grpcio==1.65.1
    # via
    #   google-api-core
    #   google-cloud-pubsub
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.65.1
    # via
    #   google-api-core
    #   google-cloud-pubsub
idna==3.7
    # via requests
oauthlib==3.2.2
    # via requests-oauthlib
packaging==24.1
    # via deprecation
proto-plus==1.24.0
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-logging
    #   google-cloud-pubsub
    #   google-cloud-secret-manager
protobuf==5.27.2
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-audit-log
    #   google-cloud-pubsub
    #   google-cloud-secret-manager
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via google-auth
requests==2.32.3
    # via
    #   google-api-core
    #   google-cloud-storage
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
urllib3==2.2.2
    # via requests

